# docs/swagger/openapi.json
{
    "components": {"schemas":{"model.Error":{"properties":{"code":{"type":"string"},"message":{"type":"string"}},"type":"object"},"model.ErrorResponse":{"properties":{"error":{"$ref":"#/components/schemas/model.Error"}},"type":"object"},"model.PrStatus":{"type":"string","x-enum-varnames":["PrStatusOpen","PrStatusMerged"]},"model.PullRequest":{"properties":{"assigned_reviewers":{"items":{"type":"string"},"type":"array","uniqueItems":false},"author_id":{"type":"string"},"merged_at":{"type":"string"},"pull_request_id":{"type":"string"},"pull_request_name":{"type":"string"},"status":{"$ref":"#/components/schemas/model.PrStatus"}},"type":"object"},"model.PullRequestShort":{"properties":{"author_id":{"type":"string"},"pull_request_id":{"type":"string"},"pull_request_name":{"type":"string"},"status":{"type":"string","x-enum-varnames":["PrStatusOpen","PrStatusMerged"]}},"type":"object"},"model.Team":{"properties":{"members":{"items":{"$ref":"#/components/schemas/model.TeamMember"},"type":"array","uniqueItems":false},"team_name":{"maxLength":64,"type":"string"}},"required":["team_name"],"type":"object"},"model.TeamMember":{"properties":{"is_active":{"type":"boolean"},"user_id":{"maxLength":32,"type":"string"},"username":{"maxLength":64,"type":"string"}},"required":["is_active","user_id","username"],"type":"object"},"model.User":{"properties":{"is_active":{"type":"boolean"},"team_name":{"type":"string"},"user_id":{"type":"string"},"username":{"type":"string"}},"type":"object"},"pull_requests.CreatePrRequest":{"properties":{"author_id":{"type":"string"},"pull_request_id":{"type":"string"},"pull_request_name":{"type":"string"}},"required":["author_id","pull_request_id","pull_request_name"],"type":"object"},"pull_requests.MergePrRequest":{"properties":{"pull_request_id":{"type":"string"}},"required":["pull_request_id"],"type":"object"},"pull_requests.PullRequestResponse":{"properties":{"pr":{"$ref":"#/components/schemas/model.PullRequest"},"replaced_by":{"type":"string"}},"type":"object"},"pull_requests.PullRequestsStatsResponse":{"properties":{"stats":{"additionalProperties":{"type":"integer"},"type":"object"}},"type":"object"},"pull_requests.ReassignReviewerPrRequest":{"properties":{"old_reviewer_id":{"type":"string"},"pull_request_id":{"type":"string"}},"required":["old_reviewer_id","pull_request_id"],"type":"object"},"teams.CreateUpdateTeamRequest":{"properties":{"members":{"items":{"$ref":"#/components/schemas/model.TeamMember"},"type":"array","uniqueItems":false},"team_name":{"maxLength":64,"type":"string"}},"required":["team_name"],"type":"object"},"users.PullRequestsAssignedToUserResponse":{"properties":{"pull_requests":{"items":{"$ref":"#/components/schemas/model.PullRequestShort"},"type":"array","uniqueItems":false},"user_id":{"type":"string"}},"type":"object"},"users.SetUserIsActiveRequest":{"properties":{"is_active":{"type":"boolean"},"user_id":{"maxLength":32,"type":"string"}},"required":["is_active","user_id"],"type":"object"}}},
    "info": {"title":"Booking Service API","version":"dev"},
    "externalDocs": {"description":"","url":""},
    "paths": {"/pullRequest/create":{"post":{"description":"Creates a new pull request with the given ID, name, and author ID.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/pull_requests.CreatePrRequest"}}},"description":"Pull request creation data","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/pull_requests.PullRequestResponse"}}},"description":"Created"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Create a new pull request","tags":["PullRequests"]}},"/pullRequest/merge":{"post":{"description":"Merges the pull request with the given ID.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/pull_requests.MergePrRequest"}}},"description":"Pull request merge data","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/pull_requests.PullRequestResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Merge a pull request","tags":["PullRequests"]}},"/pullRequest/reassign":{"post":{"description":"Reassigns the specified reviewer for the given pull request to a new reviewer.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/pull_requests.ReassignReviewerPrRequest"}}},"description":"Pull request reviewer reassignment data","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/pull_requests.PullRequestResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Reassign a reviewer for a pull request","tags":["PullRequests"]}},"/pullRequest/stats":{"get":{"description":"Retrieves statistics about pull requests.","parameters":[{"description":"Team name to filter statistics","in":"query","name":"team_name","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/pull_requests.PullRequestsStatsResponse"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Get pull requests statistics","tags":["PullRequests"]}},"/team/add":{"post":{"description":"Create a new team","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/teams.CreateUpdateTeamRequest"}}},"description":"Team to create","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.Team"}}},"description":"Created"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Add team","tags":["Teams"]}},"/team/get":{"get":{"description":"Retrieve a team by name","parameters":[{"description":"Team Name","in":"query","name":"team_name","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.Team"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Get team","tags":["Teams"]}},"/team/update":{"put":{"description":"Update an existing team","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/teams.CreateUpdateTeamRequest"}}},"description":"Team to update","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.Team"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Update team","tags":["Teams"]}},"/users/getReview":{"get":{"description":"Retrieve pull requests assigned to a specific user","parameters":[{"description":"User ID","in":"query","name":"user_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/users.PullRequestsAssignedToUserResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Get pull requests assigned to user","tags":["Users"]}},"/users/setIsActive":{"post":{"description":"Set the active status of a user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/users.SetUserIsActiveRequest"}}},"description":"Set User Is Active Request","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.User"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/model.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Set user is active","tags":["Users"]}}},
    "openapi": "3.0.1"
}